Organize tools/tasks, remove unused deps, wip update all deps
