+++ .gitignore
@@ -18,7 +18,6 @@ python
 .pydevproject
 .DS_Store
 bitly.json
-tools
 *~
 *.swp
 options
