+++ package.json
@@ -35,13 +35,13 @@
     "build": "run-s clean getcapabilities build:config build:dev",
     "build:ci": "run-s getcapabilities build:config build:prod",
     "build:options": "bash -c 'tasks/buildOptions.sh'",
-    "build:config": "run-s build:options && node ./tasks/config.js",
+    "build:config": "run-s build:options && node ./tasks/util/config.js",
     "build:dev": "cross-env NODE_ENV=development webpack",
     "build:prod": "cross-env NODE_ENV=production webpack",
-    "clean": "node ./tasks/clean.js",
+    "clean": "node ./tasks/util/clean.js",
     "dist": "run-s clean getcapabilities build:config build:prod && node ./tasks/dist.js",
     "dist:dev": "run-s clean getcapabilities build:config build:dev && node ./tasks/dist.js",
-    "distclean": "npm run clean && node ./tasks/distclean.js",
+    "distclean": "npm run clean && node ./tasks/util/distclean.js",
     "docker:ci": "npm run docker:restart && docker exec -it worldview /bin/bash /build/e2e/docker-ci.sh ; npm run docker:stop",
     "docker:image": "docker build -t worldview .",
     "docker:reimage": "docker rmi -f worldview && docker build --no-cache -t worldview .",
@@ -64,12 +64,12 @@
     "lint": "run-s lint:js lint:scss",
     "lint:scss": "npx stylelint ./web/scss/**/*.scss",
     "lint:js": "npx eslint . --quiet",
-    "preinstall": "node tasks/checkNodeVersion.js",
+    "preinstall": "node tasks/util/checkNodeVersion.js",
     "postinstall": "patch-package && npm run python",
     "python": "run-script-os",
     "python:darwin:linux": "bash -c 'tasks/pythonInstall.sh linux'",
     "python:win32": "bash -c 'tasks/pythonInstall.sh'",
-    "start": "node ./tasks/start.js",
+    "start": "node ./tasks/util/start.js",
     "test": "run-script-os",
     "test:coverage": "run-script-os",
     "test:coverage:default": "run-s test:unit:coverage test:unit:tz:coverage lint",
@@ -84,21 +84,21 @@
     "watch": "cross-env NODE_ENV=development webpack serve",
     "watch:debug:logger": "cross-env DEBUG=logger NODE_ENV=development webpack serve",
     "watch:debug:devtools": "cross-env DEBUG=devtools NODE_ENV=development webpack serve",
-    "upload": "node tasks/upload.js"
+    "upload": "node tasks/util/upload.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.10.2",
-    "@babel/eslint-parser": "^7.17.0",
-    "@babel/plugin-proposal-class-properties": "^7.10.1",
-    "@babel/preset-env": "^7.10.2",
-    "@babel/preset-react": "^7.10.1",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
-    "@webpack-cli/serve": "^1.7.0",
+    "@babel/core": "^7.20.5",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
+    "@webpack-cli/serve": "^2.0.1",
     "ajv": "^8.11.2",
-    "autoprefixer": "^10.4.12",
+    "autoprefixer": "^10.4.13",
     "axios": "^1.2.1",
-    "babel-loader": "^8.1.0",
-    "cheerio": "^1.0.0-rc.2",
+    "babel-loader": "^9.1.0",
+    "cheerio": "^1.0.0-rc.12",
     "chromedriver": "^108.0.0",
     "clean-webpack-plugin": "^4.0.0",
     "cross-env": "^7.0.2",
@@ -115,6 +115,8 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-n": "^15.6.0",
     "eslint-plugin-no-storage": "^1.0.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "^7.18.3",
     "eslint-plugin-standard": "^4.0.1",
     "express": "^4.18.1",
@@ -149,7 +151,7 @@
     "terser-webpack-plugin": "^5.3.6",
     "webpack": "^5.74.0",
     "webpack-bundle-analyzer": "^4.6.1",
-    "webpack-cli": "^4.10.0",
+    "webpack-cli": "^5.0.1",
     "webpack-dev-middleware": "^5.3.3",
     "webpack-dev-server": "^4.11.1",
     "xml-js": "^1.6.11",
@@ -184,7 +186,6 @@
     "ol-mapbox-style": "9.4.0",
     "p-queue": "^7.3.0",
     "proj4": "2.6.1",
-    "promise-queue": "2.2.5",
     "prop-types": "^15.7.2",
     "rc-slider": "9.2.2",
     "react": "^16.14.0",
@@ -212,6 +213,7 @@
     "regenerator-runtime": "^0.13.3",
     "reselect": "^4.0.0",
     "rgbcolor": "^0.0.4",
+    "simplebar": "^5.3.9",
     "simplebar-react": "^2.1.0",
     "stackblur": "^1.0.0",
     "supercluster": "7.0.0",
