+++ tasks/README.md
@@ -1,0 +1,11 @@
+# tasks
+
+A collection of tasks used in ci/cd and to build worldview locally.
+
+## bamboo
+
+## link-check
+
+## build
+
+## util
