+++ tools/README.md
@@ -1,0 +1,17 @@
+# tools
+
+A collection of "run-once" scripts.
+
+## csv2colormap
+
+## generateColormapUnitDictionary
+
+## generateLayergroup
+
+## generatePreviewImage
+
+## migrateConfigs
+
+## removeProps
+
+## renameProps
