+++ tools/csv2colormap.js
