+++ tools/generateColormapUnitDictionary.js
