+++ tools/generatePreviewImage.js
