+++ tools/migrateConfigs.js
