+++ tools/renameProps.js
