+++ web/js/app.js
@@ -36,11 +36,11 @@ import keyPress from './modules/key-press/actions';
 import setScreenInfo from './modules/screen-size/actions';
 
 // Dependency CSS
-import '../../node_modules/bootstrap/dist/css/bootstrap.min.css';
-import '../../node_modules/ol/ol.css';
-import '../../node_modules/rc-slider/dist/rc-slider.min.css';
-import '../../node_modules/simplebar/dist/simplebar.min.css';
-import '../../node_modules/react-swipe-to-delete-component/dist/swipe-to-delete.css';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import 'ol/ol.css';
+import 'rc-slider/dist/rc-slider.min.css';
+import 'simplebar/dist/simplebar.min.css';
+import 'react-swipe-to-delete-component/dist/swipe-to-delete.css';
 import 'react-image-crop/dist/ReactCrop.css';
 import 'react-resizable/css/styles.css';
 // App CSS
